resolving step added, needs proper iteration
objects being pushed into wall while they still have not yet resolved their wall collisions happens as rare as possible... there needs to be an order of precedence to make it any better

note to self: is this even worth our time? this started as an out-of-scope curiousity of the student and i rolled with it for teachable moments... but now it's gotten into "rolling our own Box2D" territory... is there anything to gain from continuing to work on this code and use it for the game jam?